pyproject.toml
[tool.poetry]
name = "eventflow-payment-service"
version = "1.0.0"
description = "EventFlow Payment Service - consumes order events and processes payments"
authors = ["EventFlow Demo"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
azure-servicebus = "^7.11.0"
httpx = "^0.26.0"
azure-monitor-opentelemetry = "^1.2.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
structlog = "^24.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
pytest-cov = "^4.1.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
